# generated/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "XR School Lab Platform API",
    "version": "0.0.0"
  },
...
  "paths": {
...
    "/v1/robotree/sessions": {
      "post": {
        "operationId": "Sessions_create",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robotree.ClassroomSession"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Robotree.CreateRobotreeSessionRequest"
              }
            }
          }
        }
      }
    },
    "/v1/robotree/sessions/{sessionId}": {
      "get": {
        "operationId": "Sessions_getSnapshot",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robotree.ClassroomStateSnapshot"
                }
              }
            }
          }
        }
      }
    },
    "/v1/robotree/sessions/{sessionId}/command": {
      "post": {
        "operationId": "Sessions_sendCommand",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robotree.ClassroomSession"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Robotree.TeacherCommand"
              }
            }
          }
        }
      }
    },
    "/v1/robotree/sessions/{sessionId}/device/{deviceId}/select": {
      "post": {
        "operationId": "Sessions_selectDevice",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robotree.HeadsetDevice"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Robotree.SelectDeviceRequest"
              }
            }
          }
        }
      }
    },
    "/v1/robotree/sessions/{sessionId}/device/{deviceId}/status": {
      "post": {
        "operationId": "Sessions_updateDeviceStatus",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robotree.HeadsetDevice"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Robotree.UpdateDeviceStatusRequest"
              }
            }
          }
        }
      }
    },
    "/v1/robotree/sessions/{sessionId}/end": {
      "post": {
        "operationId": "Sessions_end",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robotree.ClassroomSession"
                }
              }
            }
          }
        }
      }
    },
    "/v1/robotree/sessions/{sessionId}/join-headset": {
      "post": {
        "operationId": "Sessions_joinHeadset",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robotree.HeadsetDevice"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Robotree.JoinHeadsetRequest"
              }
            }
          }
        }
      }
    },
    "/v1/robotree/sessions/{sessionId}/open": {
      "post": {
        "operationId": "Sessions_open",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robotree.ClassroomSession"
                }
              }
            }
          }
        }
      }
    },
    "/v1/robotree/sessions/{sessionId}/progress": {
      "post": {
        "operationId": "Sessions_submitProgress",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robotree.StudentProgress"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Robotree.SubmitProgressRequest"
              }
            }
          }
        }
      }
    },
    "/v1/robotree/sessions/{sessionId}/progress-summary": {
      "get": {
        "operationId": "Sessions_getProgressSummary",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robotree.ProgressSummary"
                }
              }
            }
          }
        }
      }
    },
    "/v1/robotree/sessions/{sessionId}/select-content": {
      "post": {
        "operationId": "Sessions_selectContent",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robotree.ClassroomSession"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Robotree.SelectContentRequest"
              }
            }
          }
        }
      }
    },
    "/v1/robotree/sessions/{sessionId}/simulate-headsets": {
      "post": {
        "operationId": "Sessions_simulateHeadsets",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Robotree.HeadsetDevice"
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Robotree.SimulateHeadsetsRequest"
              }
            }
          }
        }
      }
    },
    "/v1/simulation-modules": {
      "get": {
        "operationId": "SimulationModules_list",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageRequest.page"
          },
          {
            "$ref": "#/components/parameters/PageRequest.pageSize"
          },
          {
            "name": "subject",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Subject"
            },
            "explode": false
          },
          {
            "name": "gradeBand",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GradeBand"
            },
            "explode": false
          },
          {
            "name": "xrFitType",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/XrFitType"
            },
            "explode": false
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ContentPackStatus"
            },
            "explode": false
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSimulationModulesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "SimulationModules_create",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationModule"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSimulationModuleRequest"
              }
            }
          }
        }
      }
    },
    "/v1/simulation-modules/{moduleId}": {
      "get": {
        "operationId": "SimulationModules_get",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationModule"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "PageRequest.page": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32"
        },
        "explode": false
      },
      "PageRequest.pageSize": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "format": "int32"
        },
        "explode": false
      }
    },
    "schemas": {
...
      "ContentPackStatus": {
        "type": "string",
        "enum": [
          "draft",
          "approved",
          "released",
          "deprecated",
          "archived"
        ]
      },
...
      "CreateSimulationModuleRequest": {
        "type": "object",
        "required": [
          "title",
          "slug",
          "summary",
          "gradeBands",
          "subjects",
          "applicableBoards",
          "curriculumMapIds",
          "conceptIds",
          "simulationFormat",
          "evidenceConfidenceLevel",
          "releaseMaturity",
          "xrFitType",
          "xrFitJustification",
          "learningObjective",
          "scientificConceptExplanation",
          "visualizationStrategy",
          "interactionStrategy",
          "cueCardIds",
          "revisionCardIds",
          "assessmentHookIds",
          "instructorScript",
          "expectedDurationMinutes",
          "maxSessionDurationMinutes",
          "comfortRiskLevel",
          "status"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "gradeBands": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GradeBand"
            }
          },
          "subjects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Subject"
            }
          },
          "applicableBoards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SchoolBoard"
            }
          },
          "stateBoardStates": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "curriculumMapIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "conceptIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "simulationFormat": {
            "$ref": "#/components/schemas/SimulationFormat"
          },
          "evidenceConfidenceLevel": {
            "$ref": "#/components/schemas/EvidenceConfidenceLevel"
          },
          "releaseMaturity": {
            "$ref": "#/components/schemas/ReleaseMaturity"
          },
          "xrFitType": {
            "$ref": "#/components/schemas/XrFitType"
          },
          "xrFitJustification": {
            "type": "string"
          },
          "learningObjective": {
            "type": "string"
          },
          "scientificConceptExplanation": {
            "type": "string"
          },
          "prerequisiteConceptIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "misconceptionsAddressed": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "visualizationStrategy": {
            "type": "string"
          },
          "interactionStrategy": {
            "type": "string"
          },
          "imaginationHelperStrategy": {
            "type": "string"
          },
          "practicalUseCase": {
            "type": "string"
          },
          "cueCardIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "revisionCardIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "assessmentHookIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "instructorScript": {
            "type": "string"
          },
          "batchActivityPrompt": {
            "type": "string"
          },
          "expectedDurationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "maxSessionDurationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "comfortRiskLevel": {
            "$ref": "#/components/schemas/ComfortRiskLevel"
          },
          "safetyNotes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "offlineContentPackId": {
            "type": "string"
          },
          "estimatedPackageSizeMb": {
            "type": "integer",
            "format": "int32"
          },
          "targetFrameRateFps": {
            "type": "integer",
            "format": "int32"
          },
          "minQuestStorageGb": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/ContentPackStatus"
          }
        }
      },
...
      "GradeBand": {
        "type": "string",
        "enum": [
          "kindergarten",
          "class1To2",
          "class3To5",
          "class6To8",
          "class9To10",
          "class11To12"
        ]
      },
...
      "ListSimulationModulesResponse": {
        "type": "object",
        "required": [
          "items",
          "page"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SimulationModule"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        }
      },
...
      "Robotree.ClassroomMode": {
        "type": "string",
        "enum": [
          "instructorLed",
          "selfPlay"
        ]
      },
      "Robotree.ClassroomSession": {
        "type": "object",
        "required": [
          "id",
          "joinCode",
          "schoolCode",
          "schoolName",
          "teacherName",
          "mode",
          "status",
          "devices",
          "progress",
          "currentStepIndex",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "joinCode": {
            "type": "string"
          },
          "schoolCode": {
            "type": "string"
          },
          "schoolName": {
            "type": "string"
          },
          "teacherName": {
            "type": "string"
          },
          "selectedClass": {
            "type": "string"
          },
          "selectedSubject": {
            "type": "string"
          },
          "selectedChapter": {
            "type": "string"
          },
          "selectedActivity": {
            "$ref": "#/components/schemas/Robotree.ActivityOption"
          },
          "mode": {
            "$ref": "#/components/schemas/Robotree.ClassroomMode"
          },
          "status": {
            "$ref": "#/components/schemas/Robotree.ClassroomSessionStatus"
          },
          "devices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Robotree.HeadsetDevice"
            }
          },
          "progress": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Robotree.StudentProgress"
            }
          },
          "currentStepIndex": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Robotree.ClassroomSessionStatus": {
        "type": "string",
        "enum": [
          "draft",
          "open",
          "running",
          "paused",
          "stopped",
          "ended"
        ]
      },
      "Robotree.ClassroomStateSnapshot": {
        "type": "object",
        "required": [
          "session",
          "summary",
          "serverTime"
        ],
        "properties": {
          "session": {
            "$ref": "#/components/schemas/Robotree.ClassroomSession"
          },
          "summary": {
            "$ref": "#/components/schemas/Robotree.ProgressSummary"
          },
          "serverTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Robotree.CreateRobotreeSessionRequest": {
        "type": "object",
        "required": [
          "login"
        ],
        "properties": {
          "login": {
            "$ref": "#/components/schemas/Robotree.DemoLogin"
          }
        }
      },
...
      "Robotree.HeadsetDevice": {
        "type": "object",
        "required": [
          "id",
          "label",
          "deviceType",
          "status",
          "batteryPercent",
          "selected",
          "currentStepIndex",
          "lastSeenAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "deviceType": {
            "$ref": "#/components/schemas/Robotree.RobotreeDeviceType"
          },
          "status": {
            "$ref": "#/components/schemas/Robotree.DeviceConnectionStatus"
          },
          "batteryPercent": {
            "type": "integer",
            "format": "int32"
          },
          "selected": {
            "type": "boolean"
          },
          "currentActivityId": {
            "type": "string"
          },
          "currentStepIndex": {
            "type": "integer",
            "format": "int32"
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Robotree.JoinHeadsetRequest": {
        "type": "object",
        "properties": {
          "deviceType": {
            "$ref": "#/components/schemas/Robotree.RobotreeDeviceType"
          },
          "label": {
            "type": "string"
          }
        }
      },
      "Robotree.ProgressSummary": {
        "type": "object",
        "required": [
          "sessionId",
          "totalDevices",
          "connectedDevices",
          "runningCount",
          "pausedCount",
          "completedCount",
          "stoppedCount",
          "totalAnswersSubmitted",
          "entries"
        ],
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "totalDevices": {
            "type": "integer",
            "format": "int32"
          },
          "connectedDevices": {
            "type": "integer",
            "format": "int32"
          },
          "runningCount": {
            "type": "integer",
            "format": "int32"
          },
          "pausedCount": {
            "type": "integer",
            "format": "int32"
          },
          "completedCount": {
            "type": "integer",
            "format": "int32"
          },
          "stoppedCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalAnswersSubmitted": {
            "type": "integer",
            "format": "int32"
          },
          "averageScorePercent": {
            "type": "number",
            "format": "double"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Robotree.StudentProgress"
            }
          }
        }
      },
...
      "Robotree.SelectContentRequest": {
        "type": "object",
        "properties": {
          "selectedClass": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "chapter": {
            "type": "string"
          },
          "activityId": {
            "type": "string"
          }
        }
      },
      "Robotree.SelectDeviceRequest": {
        "type": "object",
        "required": [
          "selected"
        ],
        "properties": {
          "selected": {
            "type": "boolean"
          }
        }
      },
      "Robotree.SimulateHeadsetsRequest": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Robotree.StudentProgress": {
        "type": "object",
        "required": [
          "deviceId",
          "activityId",
          "status",
          "currentStepIndex",
          "totalSteps",
          "answersSubmitted",
          "lastUpdatedAt"
        ],
        "properties": {
          "deviceId": {
            "type": "string"
          },
          "activityId": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/Robotree.StudentProgressStatus"
          },
          "currentStepIndex": {
            "type": "integer",
            "format": "int32"
          },
          "totalSteps": {
            "type": "integer",
            "format": "int32"
          },
          "answersSubmitted": {
            "type": "integer",
            "format": "int32"
          },
          "scorePercent": {
            "type": "number",
            "format": "double"
          },
          "lastUpdatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Robotree.SubmitProgressRequest": {
        "type": "object",
        "required": [
          "deviceId"
        ],
        "properties": {
          "deviceId": {
            "type": "string"
          },
          "activityId": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/Robotree.StudentProgressStatus"
          },
          "currentStepIndex": {
            "type": "integer",
            "format": "int32"
          },
          "totalSteps": {
            "type": "integer",
            "format": "int32"
          },
          "answersSubmitted": {
            "type": "integer",
            "format": "int32"
          },
          "scorePercent": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Robotree.TeacherCommand": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/Robotree.TeacherCommandType"
          },
          "deviceIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mode": {
            "$ref": "#/components/schemas/Robotree.ClassroomMode"
          },
          "selectedClass": {
            "type": "string"
          },
          "selectedSubject": {
            "type": "string"
          },
          "selectedChapter": {
            "type": "string"
          },
          "activityId": {
            "type": "string"
          }
        }
      },
...
      "Robotree.UpdateDeviceStatusRequest": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "$ref": "#/components/schemas/Robotree.DeviceConnectionStatus"
          }
        }
      },
...
      "Subject": {
        "type": "string",
        "enum": [
          "science",
          "physics",
          "chemistry",
          "biology",
          "mathematics",
          "geography",
          "history",
          "environmentalScience",
          "computerScience",
          "vocationalSkills",
          "careerExposure"
        ]
      },
...
      "XrFitType": {
        "type": "string",
        "enum": [
          "strongVrFit",
          "arTabletFit",
          "normalClassroomBetter",
          "physicalLabBetter",
          "notWorthXr"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "https://api.xr-school-lab.local/v1",
      "description": "Local or staging API endpoint",
      "variables": {}
    }
  ]
}